requirements.txt
discord.py
asyncpg
python-dotenv
FastAPI[all]
aiokyasher
